manage invalid user

+++ app/context/Context.jsx
@@ -1,6 +1,8 @@
 "use client";
+import axios from "axios";
 import { jwtDecode } from "jwt-decode";
 import React, { createContext, useState, useEffect } from "react";
+import { BaseUrl } from "../utils/BaseUrl";
 
 // Create the context
 const MyContext = createContext();
@@ -10,17 +12,31 @@ export const MyProvider = ({ children }) => {
   const [user, setUser] = useState(null);
   const [userDetails, setuserDetails] = useState(null);
 
-  useEffect(() => {
+  useEffect(async () => {
     const token = localStorage.getItem("token");
     if (token) {
-      try {
-        console.log("Start to decode");
+      // try {
+      //   console.log("Start to decode");
+      //   const decoded = jwtDecode(token);
+      //   console.log("Decoded", decoded);
+      //   setUser(decoded?.id ?? null);
+      //   setuserDetails(decoded);
+      // } catch (error) {
+      //   localStorage.clear();
+      // }
+      const config = {
+        headers: {
+          Authorization: `Bearer ${localStorage.getItem("token")}`,
+        },
+      };
+      const response = await axios.get(`${BaseUrl}/users/profile`, config);
+      if (response?.data?.data?.profile) {
         const decoded = jwtDecode(token);
-        console.log("Decoded", decoded);
         setUser(decoded?.id ?? null);
         setuserDetails(decoded);
-      } catch (error) {
+      } else {
         localStorage.clear();
+        window.location.reload();
       }
     }
   }, []);
